WIP added all dashboard controller

+++ src/main/java/j2ee_project/controller/catalog/category/DeleteCategoryController.java
@@ -1,6 +1,8 @@
 package j2ee_project.controller.catalog.category;
 
 import j2ee_project.Application;
+import j2ee_project.model.user.Moderator;
+import j2ee_project.model.user.TypePermission;
 import j2ee_project.service.catalog.category.CategoryService;
 import jakarta.servlet.*;
 import jakarta.servlet.http.*;
@@ -8,6 +10,8 @@ import jakarta.servlet.annotation.*;
 import org.springframework.context.ApplicationContext;
 
 import java.io.IOException;
+
+import static j2ee_project.staticServices.PermissionHelper.getPermission;
 
 /**
  * This class is a servlet used to delete a category. It's a controller in the MVC architecture of this project.
@@ -31,21 +35,29 @@ public class DeleteCategoryController extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String categoryIdStr = request.getParameter("id");
-        int categoryId = -1;
-
-        if(categoryIdStr != null && !categoryIdStr.trim().isEmpty()) {
-            try {
-                categoryId = Integer.parseInt(categoryIdStr);
-            } catch(Exception ignore) {}
-        }
-
-        if(categoryId<=0) {
-            response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Category ID must be positive");
-        }
-        categoryService.deleteCategory(categoryId);
         try {
-            response.sendRedirect("dashboard?tab=categories");
+            HttpSession session = request.getSession();
+            Object obj = session.getAttribute("user");
+            if (obj instanceof Moderator moderator
+                    && moderator.isAllowed(getPermission(TypePermission.CAN_MANAGE_CATEGORY))) {
+                String categoryIdStr = request.getParameter("id");
+                int categoryId = -1;
+
+                if (categoryIdStr != null && !categoryIdStr.trim().isEmpty()) {
+                    try {
+                        categoryId = Integer.parseInt(categoryIdStr);
+                    } catch (Exception ignore) {
+                    }
+                }
+
+                if (categoryId <= 0) {
+                    response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Category ID must be positive");
+                }
+                categoryService.deleteCategory(categoryId);
+                response.sendRedirect("dashboard?tab=categories");
+            } else {
+                response.sendRedirect("dashboard");
+            }
         }catch (Exception err){
             System.err.println(err.getMessage());
             response.sendError(HttpServletResponse.SC_NOT_FOUND);
